Add files via upload
I have the C# versions of these projects in my other repository! The only difference in this is that the tea code snippet is for tea and not coffee. If I were to go into these codes again, I think I would invite the user to create their own vibe for the tea shop, and for the Counting Sheep, maybe a prompt talking about what type of sheep you might count before sleep.